Add OSGi metadata
This adds an execution of the bnd-maven-plugin to add OSGi metadata
to the produced jar files. The metadata is slightly suboptimal in
the case of jna, because jna isn't currently exporting packages with
version numbers.  This means that the dependency on jna will work,
but it will accept more versions of jna than is desirable (you can't
say, for example, that you depend on version 4.5.1 of jna - you'll
just get all of them instead). A future version of jna should correct
this and, when that happens, the metadata here will correct itself
the next time it is generated.

This change has been tested on the Felix OSGi container and appears
to work correctly there.

See https://github.com/java-native-access/jna/issues/975